requirements.txt
# Core Flask and Extensions
Flask==3.1.0                 # Main web framework - 
Flask-Cors==5.0.0           # Handles CORS - 
Flask-SQLAlchemy==3.1.1     

# Firebase/Google Cloud
firebase-admin==6.6.0       # Firebase admin SDK - 
google-cloud-firestore==2.19.0  # Firestore database - 
google-cloud-storage==2.19.0    
google-cloud-core==2.4.1       # Core Google Cloud - 
google-auth==2.37.0           # Authentication - 
google-api-core==2.24.0       #  for Firebase
google-api-python-client==2.156.0  

# Utilities
python-dotenv==1.0.1        # For environment variables - 
blinker==1.9.0             # Flask dependency - 
certifi==2024.12.14        #  for HTTPS
charset-normalizer==3.4.0   #  for HTTP
click==8.1.7               # Flask dependency - 


flask==2.0.1
google-cloud-firestore==2.3.4
firebase-admin==5.0.3
flask-cors==3.0.10
gunicorn==20.1.0
